+++ src/gh_profiler/cli.py
@@ -18,9 +18,10 @@ from .utils.profile_data import profile_data as pdata
     is_flag=True,
     help="Generate a workflow that will automatically run `gh-profiler --concise` for every new PR and issue.",
 )
+@click.option("-v", "--verbose", is_flag=True, help="Verbose output, including explanations for decisions about flags.")
 @click.option("--redact", is_flag=True, help="Redact identifying information.")
 @click.option("--benchmark-fetch", is_flag=True, help="Benchmark the code that fetches external data.")
-def main(target, concise, generate_workflow, redact, benchmark_fetch):
+def main(target, concise, generate_workflow, verbose, redact, benchmark_fetch):
     """Examine a GitHub user's profile, to help quickly decide how much to invest in their contributions.
 
     You can target a GitHub username, or a PR/issue number from the repository you're working in.
@@ -36,10 +37,11 @@ def main(target, concise, generate_workflow, redact, benchmark_fetch):
     $ python -m gh_profiler ehmatthes
       ...
     """
-    _validate_command(target, concise, generate_workflow, redact, benchmark_fetch)
+    _validate_command(target, concise, generate_workflow, verbose, redact, benchmark_fetch)
 
     # Parse CLI options.
     pdata.concise = concise
+    pdata.verbose = verbose
     pdata.redact = redact
     pdata.benchmark_fetch = benchmark_fetch
     pdata.generate_workflow = generate_workflow
@@ -62,7 +64,7 @@ def main(target, concise, generate_workflow, redact, benchmark_fetch):
         cli_utils.process_pr_issue_num(pr_issue_num)
         gh_profiler.main()
 
-def _validate_command(target, concise, generate_workflow, redact, benchmark_fetch):
+def _validate_command(target, concise, generate_workflow, verbose, redact, benchmark_fetch):
     """Validate arguments that were passed in the CLI call."""
 
     # If target is not passed, --generate-workflow must be passed.